+++ src/filters/shaders/baseFilter.ts
@@ -1,0 +1,19 @@
+import { GLPrecision } from '../GLProbes/GLProbe';
+
+export const highPsourceCode = `precision ${GLPrecision.high} float`;
+
+export const identityFragmentShader = `
+    ${highPsourceCode};
+    varying vec2 vTexCoord;
+    uniform sampler2D uTexture;
+    void main() {
+      gl_FragColor = texture2D(uTexture, vTexCoord);
+    }`;
+
+export const vertexSource = `
+    attribute vec2 aPosition;
+    varying vec2 vTexCoord;
+    void main() {
+      vTexCoord = aPosition;
+      gl_Position = vec4(aPosition * 2.0 - 1.0, 0.0, 1.0);
+    }`;
